chore: Add impact form block
