+++ src/components/pages/blog/blog-content.js
@@ -10,6 +10,7 @@ import FootnoteContentBlock from './content-blocks/footnote-content-block'
 import RelatedPostsContentBlock from './content-blocks/related-posts-block'
 import FlippableCard from './content-blocks/flippable-card'
 import TableauChart from '~components/charts/tableau'
+import ImpactForm from '~components/common/impact-form'
 import blogContentStyles from './blog-content.module.scss'
 
 const BlogContent = ({ content }) => {
@@ -107,6 +108,9 @@ const BlogContent = ({ content }) => {
               linkToImage={imageLink}
             />
           )
+        }
+        if (__typename === 'ContentfulContentBlockImpactForm') {
+          return <ImpactForm />
         }
         if (__typename === 'ContentfulContentBlockMarkdown') {
           return (
